+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Infrastructure
 - use codecov instead of coveralls, that is failing
+- added pytest 6.2 to CI tests
 
 ## [Version 0.9.3](https://pypi.org/project/pytest-order/0.9.3/) (2021-01-14)
 Bugfix release.